requirements.txt
selenium
python-dotenv
pytz
django
psycopg2
whitenoise
pycrypto
